requirements.txt
python-telegram-bot==20.3
SQLAlchemy==2.0.19
Flask==2.3.2

